+++ src/awkward/_meta/meta.py
@@ -3,12 +3,67 @@
 from __future__ import annotations
 
 from awkward._typing import (
+    TYPE_CHECKING,
     ClassVar,
     JSONMapping,
     JSONSerializable,
     Self,
+    TypeGuard,
 )
 from awkward._util import UNSET, Sentinel
+
+if TYPE_CHECKING:
+    from awkward._meta.bitmaskedmeta import BitMaskedMeta
+    from awkward._meta.bytemaskedmeta import ByteMaskedMeta
+    from awkward._meta.indexedmeta import IndexedMeta
+    from awkward._meta.indexedoptionmeta import IndexedOptionMeta
+    from awkward._meta.listmeta import ListMeta
+    from awkward._meta.listoffsetmeta import ListOffsetMeta
+    from awkward._meta.numpymeta import NumpyMeta
+    from awkward._meta.recordmeta import RecordMeta
+    from awkward._meta.regularmeta import RegularMeta
+    from awkward._meta.unionmeta import UnionMeta
+    from awkward._meta.unmaskedmeta import UnmaskedMeta
+
+
+def is_option(
+    meta: Meta
+) -> TypeGuard[IndexedOptionMeta | BitMaskedMeta | ByteMaskedMeta | UnmaskedMeta]:
+    return meta.is_option
+
+
+def is_list(meta: Meta) -> TypeGuard[RegularMeta | ListOffsetMeta | ListMeta]:
+    return meta.is_list
+
+
+def is_numpy(meta: Meta) -> TypeGuard[NumpyMeta]:
+    return meta.is_numpy
+
+
+def is_regular(meta: Meta) -> TypeGuard[RegularMeta]:
+    return meta.is_regular
+
+
+def is_union(meta: Meta) -> TypeGuard[UnionMeta]:
+    return meta.is_union
+
+
+def is_record(meta: Meta) -> TypeGuard[RecordMeta]:
+    return meta.is_record
+
+
+def is_indexed(meta: Meta) -> TypeGuard[IndexedOptionMeta | IndexedMeta]:
+    return meta.is_indexed
+
+
+# FIXME: narrow this to have `is_tuple` be a const True
+def is_record_tuple(meta: Meta) -> TypeGuard[RecordMeta]:
+    return meta.is_record and meta.is_tuple
+
+
+# FIXME: narrow this to have `is_tuple` be a const False
+def is_record_record(meta: Meta) -> TypeGuard[RecordMeta]:
+    return meta.is_record and not meta.is_tuple
 
 
 class Meta:
